"Implement login process in Reddit scraper
Implemented the login process in the Reddit scraper. The changes include loading the login page instead of a user's saved page and automated username and password input. This change was necessary as previous scrapping was done without login, however, to access a user's saved links, a successful login is required."